Fixing some discrepancies after rebase
