+++ modules/engines/engine-hdb/src/test/resources/DuplicateTableProperties.hdbtable
